Uz mam i uzivatele

+++ src/idea/idea.service.ts
@@ -1,6 +1,6 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { IdeaEntity } from 'src/ideas/idea.entity';
+import { IdeaEntity } from 'src/idea/idea.entity';
 
 import { Repository } from 'typeorm';
 import { IdeaDTO } from './idea.dto';
@@ -24,18 +24,41 @@ export class IdeaService {
     } 
 
     async read(id: string){
-        return await this.ideaRepository.findOne({where: {id}});
+        const idea =  await this.ideaRepository.findOne({where: {id}});
+        if(!idea)
+        {
+            throw new HttpException("Not found!", HttpStatus.NOT_FOUND);
+        }
+        return idea;
     }
 
     async update(id:string, data: Partial<IdeaDTO>)
     {
+        const idea = await this.ideaRepository.findOne({where: {id}});
+        if(!idea)
+        {
+            throw new HttpException("Not found!", HttpStatus.NOT_FOUND);
+        }
         await this.ideaRepository.update({id},data);
         return await this.ideaRepository.findOne({id});
     }
 
     async destroy(id:string)
     {
+        const idea = await this.ideaRepository.findOne({where: {id}});
+        if(!idea)
+        {
+            throw new HttpException("Not found!", HttpStatus.NOT_FOUND);
+        }
         await this.ideaRepository.delete({id});
-        return {deleted: true};
+        return idea;
+    }
+
+    async deleteAll()
+    {
+        return (await this.ideaRepository.find()).forEach(element => {
+
+            this.ideaRepository.delete(element.id);
+        });
     }
 }
